refactor: move out all the unit tests and mocked unit test logic
