+++ .github/workflows/acceptance-tests-runner.yml
@@ -359,12 +359,6 @@ jobs:
           terraform_wrapper: false  
       - name: Prepare new advanced_cluster
         run: make tools enable-advancedclustertpf  
-      - name: Mocked Acceptance Tests
-        env:
-          MONGODB_ATLAS_TPF_ADV_CLUSTER_TESTS: "true"
-          ACCTEST_REGEX_RUN: '^TestMock'
-          ACCTEST_PACKAGES: ./internal/service/advancedclustertpf
-        run: make testacc
       - name: Acceptance Tests
         env:
           MONGODB_ATLAS_LAST_VERSION: ${{ needs.get-provider-version.outputs.provider_version }}
