--- internal/service/advancedclustertpf/testdata/TestMockAdvancedCluster_configSharded/01_01_POST__api_atlas_v2_groups_{groupId}_clusters_2024-08-05.json
@@ -1,40 +1,0 @@
-{
- "clusterType": "SHARDED",
- "labels": [],
- "name": "sharded-multi-replication",
- "replicationSpecs": [
-  {
-   "regionConfigs": [
-    {
-     "electableSpecs": {
-      "diskIOPS": 2000,
-      "ebsVolumeType": "PROVISIONED",
-      "instanceSize": "M30",
-      "nodeCount": 3
-     },
-     "priority": 7,
-     "providerName": "AWS",
-     "regionName": "EU_WEST_1"
-    }
-   ],
-   "zoneName": "ZoneName managed by Terraform"
-  },
-  {
-   "regionConfigs": [
-    {
-     "electableSpecs": {
-      "diskIOPS": 1000,
-      "ebsVolumeType": "PROVISIONED",
-      "instanceSize": "M30",
-      "nodeCount": 3
-     },
-     "priority": 7,
-     "providerName": "AWS",
-     "regionName": "EU_WEST_1"
-    }
-   ],
-   "zoneName": "ZoneName managed by Terraform"
-  }
- ],
- "tags": []
-}
