--- internal/service/advancedclustertpf/testdata/TestMockAdvancedCluster_configSharded/02_01_PATCH__api_atlas_v2_groups_{groupId}_clusters_{clusterName}_2024-08-05.json
@@ -1,70 +1,0 @@
-{
- "replicationSpecs": [
-  {
-   "id": "673cd6d2fb19da739ae2d4a8",
-   "regionConfigs": [
-    {
-     "analyticsSpecs": {
-      "diskIOPS": 2000,
-      "ebsVolumeType": "PROVISIONED",
-      "instanceSize": "M30",
-      "nodeCount": 1
-     },
-     "autoScaling": {
-      "compute": {
-       "enabled": false,
-       "scaleDownEnabled": false
-      },
-      "diskGB": {
-       "enabled": true
-      }
-     },
-     "electableSpecs": {
-      "diskIOPS": 2000,
-      "diskSizeGB": 40,
-      "ebsVolumeType": "PROVISIONED",
-      "instanceSize": "M30",
-      "nodeCount": 3
-     },
-     "priority": 7,
-     "providerName": "AWS",
-     "regionName": "EU_WEST_1"
-    }
-   ],
-   "zoneName": "ZoneName managed by Terraform"
-  },
-  {
-   "id": "673cd6d2fb19da739ae2d4aa",
-   "regionConfigs": [
-    {
-     "analyticsSpecs": {
-      "diskIOPS": 1000,
-      "ebsVolumeType": "PROVISIONED",
-      "instanceSize": "M30",
-      "nodeCount": 1
-     },
-     "autoScaling": {
-      "compute": {
-       "enabled": false,
-       "scaleDownEnabled": false
-      },
-      "diskGB": {
-       "enabled": true
-      }
-     },
-     "electableSpecs": {
-      "diskIOPS": 1000,
-      "diskSizeGB": 40,
-      "ebsVolumeType": "PROVISIONED",
-      "instanceSize": "M30",
-      "nodeCount": 3
-     },
-     "priority": 7,
-     "providerName": "AWS",
-     "regionName": "EU_WEST_1"
-    }
-   ],
-   "zoneName": "ZoneName managed by Terraform"
-  }
- ]
-}
