--- internal/service/advancedclustertpf/testdata/TestMockAdvancedCluster_replicaset/01_01_POST__api_atlas_v2_groups_{groupId}_clusters_2024-08-05.json
@@ -1,32 +1,0 @@
-{
- "clusterType": "REPLICASET",
- "labels": [],
- "name": "test",
- "replicationSpecs": [
-  {
-   "regionConfigs": [
-    {
-     "autoScaling": {
-      "compute": {
-       "enabled": false,
-       "scaleDownEnabled": false
-      },
-      "diskGB": {
-       "enabled": true
-      }
-     },
-     "electableSpecs": {
-      "diskSizeGB": 10,
-      "instanceSize": "M10",
-      "nodeCount": 3
-     },
-     "priority": 7,
-     "providerName": "AWS",
-     "regionName": "US_EAST_1"
-    }
-   ],
-   "zoneName": "ZoneName managed by Terraform"
-  }
- ],
- "tags": []
-}
