--- internal/service/advancedclustertpf/testdata/TestMockAdvancedCluster_replicaset/03_01_PATCH__api_atlas_v2_groups_{groupId}_clusters_{clusterName}_2024-08-05.json
@@ -1,24 +1,0 @@
-{
- "backupEnabled": true,
- "biConnector": {
-  "enabled": true,
-  "readPreference": "secondary"
- },
- "labels": [
-  {
-   "key": "env",
-   "value": "test"
-  }
- ],
- "mongoDBMajorVersion": "8.0",
- "pitEnabled": true,
- "redactClientLogData": true,
- "replicaSetScalingStrategy": "NODE_TYPE",
- "tags": [
-  {
-   "key": "env",
-   "value": "test"
-  }
- ],
- "versionReleaseSystem": "CONTINUOUS"
-}
