--- internal/service/advancedclustertpf/testdata/TestMockAdvancedCluster_replicaset/04_01_PATCH__api_atlas_v2_groups_{groupId}_clusters_{clusterName}_2024-08-05.json
@@ -1,3 +1,0 @@
-{
- "paused": false
-}
