--- internal/service/advancedclustertpf/testdata/TestMockClusterAdvancedCluster_basicTenant/01_01_POST__api_atlas_v2_groups_{groupId}_clusters_2024-08-05.json
@@ -1,21 +1,0 @@
-{
- "clusterType": "REPLICASET",
- "name": "test-acc-tf-c-91771214182147246",
- "replicationSpecs": [
-  {
-   "regionConfigs": [
-    {
-     "backingProviderName": "AWS",
-     "electableSpecs": {
-      "instanceSize": "M5",
-      "nodeCount": 0
-     },
-     "priority": 7,
-     "providerName": "TENANT",
-     "regionName": "US_EAST_1"
-    }
-   ],
-   "zoneName": "ZoneName managed by Terraform"
-  }
- ]
-}
