--- internal/service/advancedclustertpf/testdata/TestMockClusterAdvancedCluster_basicTenant/01_01_POST__api_atlas_v2_groups_{groupId}_clusters_2024-08-05_manual.json
@@ -1,22 +1,0 @@
-{
- "clusterType": "REPLICASET",
- "labels": [],
- "name": "test-acc-tf-c-8049930413007488732",
- "replicationSpecs": [
-  {
-   "regionConfigs": [
-    {
-     "backingProviderName": "AWS",
-     "electableSpecs": {
-      "instanceSize": "M5"
-     },
-     "priority": 7,
-     "providerName": "TENANT",
-     "regionName": "US_EAST_1"
-    }
-   ],
-   "zoneName": "ZoneName managed by Terraform"
-  }
- ],
- "tags": []
-}
