--- internal/testutil/unit/testdata/TestMockRoundTripper/03_01_POST__api_atlas_v2_orgs_{orgId}_resourcePolicies:validate_2024-08-05.json
@@ -1,8 +1,0 @@
-{
- "name": "updated-policy",
- "policies": [
-  {
-   "body": "\t\t\t\n\tforbid (\n\tprincipal,\n\taction == cloud::Action::\"cluster.createEdit\",\n\tresource\n\t) when {\n\tcontext.cluster.cloudProviders.containsAny([cloud::cloudProvider::\"aws\"])\n\t};\n"
-  }
- ]
-}
