--- internal/testutil/unit/testdata/TestMockRoundTripper/03_01_POST__api_atlas_v2_orgs_{orgId}_resourcePolicies:validate_2024-08-05_manual.json
@@ -1,3 +1,0 @@
-{
- "name": "overriding the filename by using _manual"
-}
